Split tests for python3 in all and external_modules
