+++ flavors/python-3.6-data-science-EXASOL-6.2.0/flavor_base/testconfig
@@ -1,2 +1,2 @@
 generic_language_tests=python3
-test_folders=python3 python3-ds-flavor
+test_folders=python3/all python3/external_modules python3-ds-flavor
