+++ flavors/python-3.6-minimal-EXASOL-6.2.0/flavor_base/testconfig
@@ -1,2 +1,2 @@
 generic_language_tests=python3
-test_folders=python3
+test_folders=python3/all
