+++ tests/test/python3/all/env_directive.py
