+++ tests/test/python3/all/general.py
@@ -143,29 +143,6 @@ class PythonImport(udf.TestCase):
         self.query(sql % {'module': module})
         self.query('SELECT i() FROM DUAL')
 
-    modules = '''
-            ujson
-            lxml
-            numpy
-            pytz
-            scipy
-            '''.split()
-    @useData((x,) for x in modules)
-    def test_3rdparty_modules_are_importable(self, module):
-        sql = udf.fixindent('''
-            CREATE OR REPLACE python3 SCALAR SCRIPT
-            i()
-            RETURNS VARCHAR(1024) AS
-            import %(module)s
-            def run(ctx):
-                try:
-                    return %(module)s.__file__
-                except AttributeError:
-                    return "(built-in)"
-            ''')
-        self.query(sql % {'module': module})
-        self.query('SELECT i() FROM DUAL')
-
     # modules = '''
     #         '''.split()
     # @useData((x,) for x in modules)
