+++ tests/test/python3/all/script_import.py
